chore: create distance functions file
